Sandco/mysite/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-recaptcha==4.0.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2

django-modeltranslation~=0.19.11